add a tool for file cleanups

+++ buildtools/clearUi.py
@@ -3,17 +3,31 @@ import re
 
 
 
-def removeAllUiFiles():
+def removeAllUicFiles():
     files_dir = os.path.join(os.getcwd(), 'ui')
     files = os.listdir(files_dir)
     to_remove = []
-    pat = re.compile('ui_*.py')
+    pat = re.compile(r'^ui_.*\.py')
     for f in files:
-        if re.match(pat):
-            to_remove.add(f)
+        if pat.match(f):
+            to_remove.append(f)
+
+    if len(to_remove) == 0:
+        print('no uic files found at: ' + files_dir)
+        return
+
+    print('files to remove:')
     for f in to_remove:
-        os.remove(os.path.join(files_dir, f))
+        print(f)
+    
+    option = input('remove? [y/n]')
+    if option == 'y':
+        for f in to_remove:
+            os.remove(os.path.join(files_dir, f))
+        print('uic files removed')
+    else:
+        print('remove canceled')
 
 
 if __name__ == '__main__':
-    removeAllUiFiles()
+    removeAllUicFiles()
